fix: Reorganize DefaultBanksComponent test structure
Remove obsolete test files and consolidate into single
DefaultBanksComponentTests with improved organization, descriptive
naming, and clear MARK sections.
